refactor: setup refactor
