--- apps/web/src/features/setup/components/SetupLoadingScreen.stories.tsx
@@ -1,14 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import { SetupLoadingScreen } from './SetupLoadingScreen';
-
-type Story = StoryObj<typeof SetupLoadingScreen>;
-
-export default {
-  component: SetupLoadingScreen,
-  parameters: {
-    layout: 'fullscreen'
-  }
-} as Meta<typeof SetupLoadingScreen>;
-
-export const Default: Story = {};
