+++ apps/web/src/features/setup/pages/SetupLoadingPage/SetupLoadingPage.stories.tsx
@@ -1,0 +1,14 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import { SetupLoadingPage } from './SetupLoadingPage';
+
+type Story = StoryObj<typeof SetupLoadingPage>;
+
+export default {
+  component: SetupLoadingPage,
+  parameters: {
+    layout: 'fullscreen'
+  }
+} as Meta<typeof SetupLoadingPage>;
+
+export const Default: Story = {};
