+++ apps/web/src/features/setup/pages/SetupLoadingPage/index.ts
@@ -1,0 +1,1 @@
+export * from './SetupLoadingPage';
