+++ apps/web/src/features/setup/pages/SetupPage/index.ts
@@ -1,0 +1,1 @@
+export * from './SetupPage';
